Reorganized the Header into 2 seperate components to combat file size bloat.
